Elektra: mv rand to extra file + remove rand_r dep + opmphm for all
